requirements.txt
streamlit==1.28.1
openai==1.3.5
pinecone-client==2.2.4
langchain==0.0.330
langchain-openai==0.0.2
PyMuPDF==1.23.8
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
tiktoken==0.5.1
streamlit-chat==0.1.1